feat: support deploy and update command

+++ packages/harsta/node/cli/compile.ts
@@ -70,18 +70,23 @@ export function registerCompileCommand(cli: Argv) {
       ])
 
       await buildTypes()
+      await buildDistributed()
 
-      const defaultOutput = clientRoot ? path.resolve(clientRoot, 'build') : path.resolve(userRoot, 'dist')
-      const output = args.output ? (path.isAbsolute(args.output) ? args.output : path.resolve(userRoot, args.output)) : undefined
-      const outdir = output || defaultOutput
+      async function buildDistributed() {
+        if (args.output === 'ONLY_COMPILE')
+          return
+        const defaultOutput = clientRoot ? path.resolve(clientRoot, 'build') : path.resolve(userRoot, 'dist')
+        const output = args.output ? (path.isAbsolute(args.output) ? args.output : path.resolve(userRoot, args.output)) : undefined
+        const outdir = output || defaultOutput
 
-      args.clean && await fs.remove(outdir)
+        args.clean && await fs.remove(outdir)
 
-      exec(`node ${tsupBinRoot} --outDir ${outdir}`, { stdio: 'ignore', cwd: generateRoot })
-      exec(`node ${tscBinRoot} --outDir ${outdir}`, { cwd: generateRoot })
+        exec(`node ${tsupBinRoot} --outDir ${outdir}`, { stdio: 'ignore', cwd: generateRoot })
+        exec(`node ${tscBinRoot} --outDir ${outdir}`, { cwd: generateRoot })
 
-      const log = path.resolve(outdir, '../').endsWith('@harsta/client') ? '@harsta/client' : outdir
-      consola.log(`\n✔ Generated Harsta Client ${dim(`to ${log}`)}\n`)
+        const log = path.resolve(outdir, '../').endsWith('@harsta/client') ? '@harsta/client' : outdir
+        consola.log(`\n✔ Generated Harsta Client ${dim(`to ${log}`)}\n`)
+      }
 
       async function buildAddresses() {
         const packAddressesPath = path.resolve(generateRoot, './addresses/index.ts')
